POD support: Use ns-listerner pid as containers' pid
For the sake of K8S integration, we must have a mechanism to create a
pod and allow to integrate with CNI network plugins.

As CNI plugins are operating on nslistener namespaces, we should set pid
of nslistener as container's init pid, so that when we insert a NIC into
the "fake" namespace, the VM can get notifications.

Signed-off-by: Zhang Wei <zhangwei555@huawei.com>