pass multiple mountpoint of same device to hypervisor
make volume mountpoint in a container a slice so that we can
save multiple mountpoints to it rather than assuming there is
only one mountpoint of same device in a container.

Do the same to volume readOnly field.

With both above, we create multiple entries in a container's fsmap
when a volume is mounted at different mountpoints.

Signed-off-by: Peng Tao <bergwolf@gmail.com>