read until EOF in streamCopy
If either stdoutPipe or stderrPipe reads to EOF, do not close the other
one so that it can read to EOF as well. Otherwise we might lose output
in the other pipe.

This works because if hyperstart ends the stdout/stderr streams, it always
closes the writer part of both pipes. Then io.Copy() will get EOF and
return success here.

Also no need to wait stdin go routine because streamCopy() runs in a go
routine itself thus no one is really waiting out there.

Signed-off-by: Peng Tao <bergwolf@gmail.com>