shim: run shim in specified netns if necessary
In docker case, we let the shim to own a netns and for shared netns
containers, we let new shims join existing netns.

We used to use nsenter to do it for runv shim, but kata shim does not
support nsenter. So we change to handle it when spawning new shims and
it works for both runv shim and kata shim.

Signed-off-by: Peng Tao <bergwolf@gmail.com>